Use hof-theme-govuk@2
Because `hof` loads the latest theme by default if none is specified in the parent project then it is combining the old css from `hof-frontend-toolkit` and the new html from the theme at version 3.

By specifying the theme at version 2 then we ensure the html and css remain in sync.