Upgrade to hof 13 &  hof-theme-govuk 4.0.0 which fixes the issue of session timeout page not going back to start
hof 13 upgrades hof-middleware 2.0.0: locals startLink to the baseUrl or / if there is no baseUrl. This is used in the error template

The error template is found in hof-template-partials which is a dependency of hof-theme-govuk NOT hof . hof-template-partials has also been updated to 4.0.0. Remove `/` from startlink of the template, it'll be passed in by locals from hof-middleware.