[LCS-381] Fix empty application pdfs (#107)
* Split summary page behaviour to share with pdf generation

The existing `confirm-tenants` behaviour was handling multiple separate concerns - particularly the rendering of tenant data out to the table, the handling of outbound edit links, and the reconcilation of tenant data when returning from an edit journey.

The first of these needs to be re-used for the data pdf sent on submission, so split that into a separate behaviour so it can be reused across both pages.

* Fix pdf template

The data structure generated by the pdf behaviour does not match that represented in the template. Fix to use the same html as the summary page and so correctly match the data structure.

* Include app css in thtml to be rendered to pdf

This allows the pdf to use the app's default styles without needing to manually re-create them.

* Include summary sections behaviour in pdf uploader

Since the pdf uploader cannot work properly without extending the summary behaviour, it is better to have it explicitly require it internally.