Fix date of birth being removed from tenants array (#112)
Code that mapped fields from top level of model to tenants array was expecting fields to always be defined as strings. Because date of birth includes a parse function it was defined as an object so was not picked up.

Add a `map` to normalise field configuration to list of field keys.