[LCS-403] - Amend start content markdown
Now that hof is using showdown, it appears that the markdown was not displaying fully. I had to rewrite the bold part for it to display.  It appears that the current text has something that is not making it display

Also two spaces are required for a line break md files